+++ src/RuleSet.php
@@ -70,6 +70,19 @@ final class RuleSet
     }
 
     /**
+     * Returns a new rule set with merged rules.
+     */
+    public function withRules(Rules $rules): self
+    {
+        return new self(
+            $this->customFixers,
+            $this->name,
+            $this->phpVersion,
+            $this->rules->merge($rules),
+        );
+    }
+
+    /**
      * Returns a new rule set with rules where the header_comment fixer is enabled to add a header.
      *
      * @see https://github.com/PHP-CS-Fixer/PHP-CS-Fixer/blob/v3.27.0/doc/rules/comment/header_comment.rst